% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncomplyR.r
\docType{data}
\name{vitaminA}
\alias{vitaminA}
\title{Vitamin A Randomized Trial Data Set}
\format{A data frame with 23682 rows and 3 variables}
\usage{
data(vitaminA)
}
\description{
A dataset containing the results of an RCT investigating the effect
of Vitamin A supplementation on mortality. The variables are as follows:
}
\details{
\itemize{
  \item survived. indicator of survival (survived = 1) or death (survived = 0)
  \item vitaminA_assigned. indicator of assignment to Vitamin A (vitaminA_assigned = 1) or
  placebo (vitaminA_assigned = 0)
  \item vitaminA_received. indicator of whether Vitamin A supplements were received (vitaminA_received = 1)
  or not received (vitaminA_received = 0)
}
}
\examples{
\dontrun{vitaminA}
}
\keyword{datasets}

